+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/data/validators/MaritalStatusCodeValidator.java
@@ -30,7 +30,7 @@ public class MaritalStatusCodeValidator implements ConstraintValidator<MaritalSt
         if (value == null || value.getClass().isAssignableFrom(String.class)) {
             return value == null || StringUtils.isBlank(value.toString()) || maritalStatusLookupService.existsByCode(value.toString().toUpperCase());
         } else {
-            return EMaritalStatus.fromId((Long) value) != null;
+            return EMaritalStatus.fromId((Integer) value) != null;
         }
     }
